Clean up install.
